Move level detection to separate file to reduce code complexity
